+++ app.tf
@@ -5,13 +5,15 @@ data "ns_app_env" "this" {
 }
 
 locals {
-  app_namespace = "default"
+  app_namespace = local.kubernetes_namespace
   app_name      = data.ns_workspace.this.block_name
   app_version   = coalesce(data.ns_app_env.this.version, "latest")
 }
 
 locals {
   app_metadata = tomap({
-    // Inject app metadata into capabilities here (e.g. security_group_name, role_name)
+    // Inject app metadata into capabilities here (e.g. service_account_id)
+    service_account_id    = google_service_account.app.id
+    service_account_email = google_service_account.app.email
   })
 }