+++ gcp.tf
@@ -1,0 +1,7 @@
+data "google_client_config" "this" {}
+
+locals {
+  region     = data.google_client_config.this.region
+  zone       = data.google_client_config.this.zone
+  project_id = data.google_client_config.this.project
+}